Fix linter errors

+++ app/helpers/sessions_helper.rb
@@ -13,9 +13,7 @@ module SessionsHelper
 
   def logout_link
     content = ''
-    if logged_in?
-      content << link_to('Logout', logout_path, method: 'delete', class: 'nav-link')
-    end
+    content << link_to('Logout', logout_path, method: 'delete', class: 'nav-link') if logged_in?
     content.html_safe
   end
 end
